Refactor: User 도메인 query 중앙화
