+++ src/app/user/[id]/activity/page.tsx
@@ -1,7 +1,5 @@
 import { notFound } from "next/navigation";
 import { ActivitySection } from "@/components/domains/user/sections";
-import { userApi } from "@/lib/utils/api/users.api";
-import { ApiMyProfileResponse } from "@/lib/hono/schemas/me.schema";
 import { getUserSessionInfo } from "@/lib/utils/getUserSessionInfo";
 
 export default async function ActivityPage({
@@ -10,20 +8,7 @@ export default async function ActivityPage({
   params: Promise<{ id: string }>;
 }) {
   const { id: pageOwnerId } = await params;
-
   const { isMe } = await getUserSessionInfo(pageOwnerId);
-
-  let user: ApiMyProfileResponse | null;
-
-  try {
-    user = await userApi.getById(pageOwnerId);
-    if (!user) {
-      return notFound();
-    }
-  } catch (error) {
-    console.error("Error fetching user:", error);
-    return notFound();
-  }
 
   if (!isMe) {
     return notFound();
