+++ src/app/user/[id]/layout.tsx
@@ -3,7 +3,7 @@ import getQueryClient from "@/lib/utils/getQueryClient";
 import { userApi } from "@/lib/utils/api/users.api";
 import UserLayoutClient from "@/components/domains/user/UserLayoutClient";
 import { getUserSessionInfo } from "@/lib/utils/getUserSessionInfo";
-import { notFound } from "next/navigation";
+import { queryKeys } from "@/lib/utils/queryKeys";
 
 interface LayoutProps {
   children: React.ReactNode;
@@ -15,15 +15,10 @@ export default async function UserLayout({ children, params }: LayoutProps) {
   const queryClient = getQueryClient();
   const { isMe } = await getUserSessionInfo(pageOwnerId);
 
-  try {
-    await queryClient.prefetchQuery({
-      queryKey: ["user", "profile", pageOwnerId],
-      queryFn: () => userApi.getById(pageOwnerId),
-    });
-  } catch (error) {
-    console.error("Error fetching user:", error);
-    return notFound();
-  }
+  await queryClient.prefetchQuery({
+    queryKey: queryKeys.user.profile(pageOwnerId),
+    queryFn: () => userApi.getById(pageOwnerId),
+  });
 
   return (
     <HydrationBoundary state={dehydrate(queryClient)}>
