+++ src/components/domains/user/UserLayoutClient.tsx
@@ -3,7 +3,7 @@
 import UserHeader from "@/components/domains/user/sections/ProfileHeaderSection";
 import UserFooter from "@/components/domains/user/sections/RecentViewSection";
 import PageClient from "@/components/domains/user/PageClient";
-import { useUserProfileById } from "@/lib/hooks/useUserProfile";
+import { useUserProfileById } from "@/lib/hooks/query/useUserQuery";
 import { notFound } from "next/navigation";
 
 interface UserLayoutClientProps {
