+++ src/components/domains/user/sections/BookmarkSection/loader/CommunityBookmarksLoader.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { CommunityBookmarkList } from "../components";
-import { useUserPostsBookmarks } from "../hooks/useUserBookmarks";
+import { useUserPostsBookmarks } from "@/lib/hooks/query/useUserQuery";
 
 export default function CommunityBookmarksLoader() {
   const { data: communityData } = useUserPostsBookmarks();
