+++ src/components/domains/user/sections/ProfileSection/components/ProfileImage/ActionButtons.tsx
@@ -1,3 +1,6 @@
+import { ButtonFilledPrimaryLFit } from "@/components/commons/button/ButtonFilled";
+import { ButtonOutlinedPrimaryLFit } from "@/components/commons/button/ButtonOutlined";
+
 interface ActionButtonsProps {
   isUploading: boolean;
   onSave: () => void;
@@ -11,22 +14,12 @@ export const ActionButtons = ({
 }: ActionButtonsProps) => {
   return (
     <div className="flex gap-2 mt-2">
-      <button
-        type="button"
-        onClick={onSave}
-        disabled={isUploading}
-        className="px-4 py-2 text-sm font-medium text-white bg-indigo-600 border border-transparent rounded-md shadow-sm hover:bg-indigo-700 disabled:bg-indigo-300"
-      >
+      <ButtonFilledPrimaryLFit onClick={onSave} disabled={isUploading}>
         {isUploading ? "저장 중..." : "이미지 저장"}
-      </button>
-      <button
-        type="button"
-        onClick={onCancel}
-        disabled={isUploading}
-        className="px-4 py-2 text-sm font-medium text-gray-700 bg-white border border-gray-300 rounded-md shadow-sm hover:bg-gray-50"
-      >
+      </ButtonFilledPrimaryLFit>
+      <ButtonOutlinedPrimaryLFit onClick={onCancel} disabled={isUploading}>
         취소
-      </button>
+      </ButtonOutlinedPrimaryLFit>
     </div>
   );
 };
