+++ src/components/domains/user/sections/ProfileSection/components/ProfileImage/ImageDisplay.tsx
@@ -1,4 +1,5 @@
 import Image from "next/image";
+import ICONS from "@/lib/constants/icons";
 interface ImageDisplayProps {
   imageUrl?: string | null;
   previewUrl?: string | null;
@@ -32,22 +33,14 @@ export const ImageDisplay = ({
       <button
         type="button"
         onClick={onEditClick}
-        className="absolute bottom-1 right-1 flex items-center justify-center w-10 h-10 bg-slate-700 rounded-full border-2 border-white shadow-md hover:bg-slate-800 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 transition-colors"
+        className="absolute bottom-1 right-1 flex items-center justify-center w-10 h-10 bg-white rounded-full border-2 border-primary-100 shadow-md hover:bg-primary-600 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-primary-300 transition-colors"
         aria-label="프로필 사진 변경"
       >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          className="h-5 w-5 text-white"
-          viewBox="0 0 20 20"
-          fill="currentColor"
-        >
-          <path d="M17.414 2.586a2 2 0 00-2.828 0L7 10.172V13h2.828l7.586-7.586a2 2 0 000-2.828z" />
-          <path
-            fillRule="evenodd"
-            d="M2 6a2 2 0 012-2h4a1 1 0 010 2H4v10h10v-4a1 1 0 112 0v4a2 2 0 01-2 2H4a2 2 0 01-2-2V6z"
-            clipRule="evenodd"
-          />
-        </svg>
+        <Image
+          src={ICONS.Insta.src}
+          alt="프로필 사진 변경"
+          className="[filter:brightness(0)_saturate(100%)_invert(28%)_sepia(14%)_saturate(1140%)_hue-rotate(328deg)_brightness(95%)_contrast(89%)]"
+        />
       </button>
     </div>
   );
