+++ src/lib/constants/icons.ts
@@ -22,6 +22,7 @@ import {
   ArrowRightGray,
   GlobePrimary,
   Reply,
+  Insta,
   DislikeGray,
   DislikePrimary,
   BadGray,
@@ -71,6 +72,7 @@ const ICONS: { [key: string]: TIcon } = {
   View: { src: View, alt: "view" },
   CloseCircle: { src: CloseCircle, alt: "close" },
   Reply: { src: Reply, alt: "reply" },
+  Insta: { src: Insta, alt: "insta" },
   DislikeGray: { src: DislikeGray, alt: "dislike" },
   DislikePrimary: { src: DislikePrimary, alt: "dislike" },
   BadGray: { src: BadGray, alt: "bad" },
